More progress on testing
There is now a test that runs the backup script and verifies an archive
was created with the expected contents.
And a test that runs the restore script which verifies the expected
contents were restored.

This is still a WIP but getting better.
Other parts of the bash script need to set -C to tar correctly to the
path parsed from the url.

Need to figure out how to test ssh connections in a clean env.